Fixed Bullet Firing
Fixed a bug in the player class where the scale would cause it to be
drawn at incorrect screen positions.

+++ Glitch/GameVariables.cs
@@ -15,8 +15,8 @@ namespace Glitch
     static class GameVariables
     {
         //static game attributes
-        public static int WINDOW_WIDTH;
-        public static int WINDOW_HEIGHT;
+        public static float WINDOW_WIDTH;
+        public static float WINDOW_HEIGHT;
         public static int NUMBER_OF_ROOMS;
         public static int NUMBER_OF_ENEMIES;
         public static int DENSITY_OF_TRAPS;
